You can now query rooms.
The percent issue still needs addressing.
