+++ src/room_util.h
@@ -41,6 +41,7 @@ TbBool delete_room_slab(MapSlabCoord slb_x, MapSlabCoord slb_y, unsigned char gn
 TbBool replace_slab_from_script(MapSlabCoord slb_x, MapSlabCoord slb_y, unsigned char slabkind);
 short check_and_asimilate_thing_by_room(struct Thing *thing);
 EventIndex update_cannot_find_room_wth_spare_capacity_event(PlayerNumber plyr_idx, struct Thing *creatng, RoomKind rkind);
+void query_room(struct Room *room);
 /******************************************************************************/
 #ifdef __cplusplus
 }
