+++ src/thing_data.c
@@ -23,7 +23,7 @@
 #include "bflib_sound.h"
 #include "bflib_memory.h"
 #include "bflib_math.h"
-
+#include "frontend.h"
 #include "config_creature.h"
 #include "config_effects.h"
 #include "thing_stats.h"
@@ -263,6 +263,47 @@ void set_thing_draw(struct Thing *thing, long anim, long speed, long scale, char
       thing->field_48 = i;
       thing->field_40 = i << 8;
     }
+}
+
+void query_thing(struct Thing *thing)
+{
+    const char title[24];
+    const char* name = thing_model_name(thing);
+    const char owner[24]; 
+    const char health[24];
+    const char position[24];
+    const char amount[24] = "\0";
+    sprintf(title, "Thing ID: %d", thing->index);
+    sprintf(owner, "Owner: %d", thing->owner);
+    sprintf(position, "Pos: X:%d Y:%d Z:%d", thing->mappos.x.stl.num, thing->mappos.y.stl.num, thing->mappos.z.stl.num);
+    if (thing->class_id == TCls_Trap)
+    {
+        struct ManfctrConfig *mconf = &gameadd.traps_config[thing->model];
+        sprintf(health, "Shots: %d/%d", thing->trap.num_shots, mconf->shots);
+    }
+    else
+    {
+        if (thing->class_id == TCls_Object)
+        {
+            if (object_is_gold(thing))
+            {
+                sprintf(amount, "Amount: %d", thing->valuable.gold_stored);   
+            }
+        }  
+        sprintf(health, "Health: %d", thing->health);
+        if (thing->class_id == TCls_Door)
+        {
+            sprintf(health, "%s/%d", health, door_stats[thing->model][0].health);
+        }
+        else if (thing->class_id == TCls_Object)
+        {
+            if (thing->model == 5)
+            {
+                sprintf(health, "%s/%d", health, game.dungeon_heart_health);
+            }
+        }
+    }
+    create_message_box(&title, name, &owner, &health, &position, &amount);
 }
 /******************************************************************************/
 #ifdef __cplusplus
