+++ src/thing_data.h
@@ -262,6 +262,8 @@ TbBool thing_is_dragged_or_pulled(const struct Thing *thing);
 struct PlayerInfo *get_player_thing_is_controlled_by(const struct Thing *thing);
 
 void set_thing_draw(struct Thing *thing, long anim, long speed, long scale, char a5, char start_frame, unsigned char draw_class);
+
+void query_thing(struct Thing *thing);
 /******************************************************************************/
 #ifdef __cplusplus
 }
